requirements.txt
transformers>=4.35.0
coremltools>=8.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
sentencepiece>=0.1.99
